LayoutTests/crypto/subtle/deriveBits-malformed-parameters-expected.txt

Test deriveBits operation with malformed parameters

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS crypto.subtle.deriveBits() rejected promise  with TypeError: Not enough arguments.
PASS crypto.subtle.deriveBits(1) rejected promise  with TypeError: Not enough arguments.
PASS crypto.subtle.deriveBits(1, 2) rejected promise  with TypeError: Not enough arguments.
PASS crypto.subtle.deriveBits("ECDH", 2, 3) rejected promise  with NotSupportedError (DOM Exception 9): The operation is not supported..
PASS successfullyParsed is true

TEST COMPLETE
